requirements.txt
redis>=5.0,<6.0
elasticsearch>=8.0,<9.0
httpx>=0.27,<1.0
pydantic>=2.0,<3.0
pyyaml>=6.0,<7.0
structlog>=24.0
tenacity>=8.0,<10.0
